# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eunomia API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "1"
  },
  "paths": {
...
    "/api/chat/threads": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List chat threads",
        "operationId": "list_threads",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThreadOut"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create a chat thread",
        "operationId": "create_thread",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadOut"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/chat/threads/{thread_id}": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send a message, streamed back as server-sent events",
        "operationId": "send_message",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Server-sent events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "Delete a chat thread",
        "operationId": "delete_thread_route",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkBody"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/chat/threads/{thread_id}/history": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Messages in a thread",
        "operationId": "thread_history",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageOut"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "MessageOut": {
        "type": "object",
        "required": [
          "role",
          "content",
          "thread_id",
          "created_at"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "thread_id": {
            "type": "string"
          },
          "tool_calls": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "object"
            }
          }
        }
      },
      "OkBody": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
      "Problem": {
        "type": "object",
        "description": "RFC 9457 `application/problem+json`, as built in `error.rs`.",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable dotted error code, see `docs/errors.md`."
          },
          "detail": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "The HTTP reason phrase."
          },
          "trace_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "type": "string",
            "description": "Always `about:blank`."
          }
        }
      },
...
      "ThreadCreate": {
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ThreadOut": {
        "type": "object",
        "required": [
          "id",
          "title",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "eunomia_session"
      }
    }
  }
}